Computes the DSE decomposition and DVARS-related statistics. Citation: Insight and inference for DVARS (Afyouni and Nichols, 2018)

Differences from implementation at github.com/asoroosh/DVARS:

The matrix is transposed.

We center and scale the matrix differently (see scale_med)

We set all zero-variance voxels to zero during centering & scaling. This means that when we remove constant 0 or NA voxels, constant non-zero voxels are also removed.

We use a tolerance of \(1e-8\) to detect non-zero voxels.

DVARS(X, normalize = FALSE, norm_I = 100, verbose = FALSE)

Arguments

X
a T x N numeric matrix representing an fMRI run.

normalize
Normalize the data as proposed in the original paper? Default is FALSE.

norm_I
The value to scale to. Default is 100, as in the original paper.

verbose
Should occasional updates be printed? Default is FALSE.