CompCor

The CompCor algorithm (Behzadi et. al., 2007) 10.1016/j.neuroimage.2007.04.042

CompCor(
  vol,
  data_ROI = NULL,
  data = NULL,
  noise_ROI,
  noise_nPC = 5,
  noise_erosion = 0
)

Arguments

vol
The \(I x J x K x T\) volumetric fMRI timeseries, including both the data voxels (unless data is provided) and the noise voxels.

data_ROI
A logical mask for vol where TRUE values indicate data voxels, from which the noise components will be regressed. Up to one of data_ROI or data should be provided.

data
A \(V x T\) data matrix from which the noise components will be regressed. Up to one of data_ROI or data should be provided.

noise_ROI
A list of logical masks for vol where TRUE values indicate voxels in a certain "noise ROI" (e.g. the WM or CSF). Noise components will be obtained separately for each noise ROI.

noise_nPC
A vector the same length as noise_ROI indicating the number of PCs from each noise ROI to use. If this vector is a shorter length its elements will be recycled. The values can also be between 0 and 1, in which case they will represent the minimum proportion of variance explained by the PCs used for each noise ROI. The smallest number of PCs will be used to achieve this proportion of variance explained. Default: 5.

noise_erosion
The number of layers to erode from each of the noise_ROI. Default: 0.

Value

A list with entries "data", "noise", "noise_mask" and "noise_var".

If data_ROI or data was provided, the entry "data" will be a V x T matrix where each row is a data voxel (if data_ROI was provided, the voxels are in spatial order; if data was provided, the voxels are in the same order) time series with each noise PC regressed from it. Otherwise, this entry will be NULL.

The entry "noise" is a list of T \(x\) noise_nPC PC scores, one for each noise_ROI.

Details

The data are centered on each voxel timecourse's median.